requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bidict==0.22.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.101.1
h11==0.14.0
httptools==0.6.0
idna==3.4
pydantic==2.2.1
pydantic_core==2.6.1
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
